requirements.txt
pyfiglet~=0.8.post1
spotipy~=2.19.0
python-dotenv~=0.19.2
ytmusicapi~=0.20.0